+++ tests/TestModel.php
@@ -2,12 +2,12 @@
 
 namespace Spatie\MediaLibrary\Test;
 
+use Spatie\MediaLibrary\HasMedia\Interfaces\HasMedia;
 use Spatie\MediaLibrary\Media;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\MediaLibrary\HasMedia\HasMediaTrait;
-use Spatie\MediaLibrary\HasMedia\Interfaces\HasMediaConversions;
 
-class TestModel extends Model implements HasMediaConversions
+class TestModel extends Model implements HasMedia
 {
     use HasMediaTrait;